upload by uploadify

+++ trunk/src/main/java/com/shw/netdisk/storage/FileSystemStorageService.java
@@ -1,5 +1,14 @@
 package com.shw.netdisk.storage;
 
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.nio.file.FileAlreadyExistsException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.List;
+import java.util.stream.Stream;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
@@ -7,18 +16,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.FileSystemUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.nio.file.FileAlreadyExistsException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.stream.Stream;
+import com.shw.netdisk.service.FileService;
 
 @Service
 public class FileSystemStorageService implements StorageService {
 
     private final Path rootLocation;
+    
+    @Autowired
+    private FileService fileService;
 
     @Autowired
     public FileSystemStorageService(StorageProperties properties) {
@@ -32,7 +38,12 @@ public class FileSystemStorageService implements StorageService {
             if (file.isEmpty()) {
                 throw new StorageException("Failed to store empty file " + file.getOriginalFilename());
             }
-            Files.copy(file.getInputStream(), this.rootLocation.resolve(file.getOriginalFilename()));
+            if(this.existed(targetPath)){
+            	String newFileName = refactorFileName(file);
+            	Files.copy(file.getInputStream(), this.rootLocation.resolve(newFileName));
+            }else{
+            	Files.copy(file.getInputStream(), this.rootLocation.resolve(file.getOriginalFilename()));
+            }
         } catch (FileAlreadyExistsException e) {
 			System.out.println("file is existed");
 		}catch (IOException e) {
@@ -40,6 +51,20 @@ public class FileSystemStorageService implements StorageService {
         }
         return targetPath;
     }
+    
+    private String refactorFileName(MultipartFile file) {
+    	String fileName = file.getOriginalFilename();
+    	List<String> likeFileNames = fileService.listByName(fileName);
+    	String lastFileName = likeFileNames.get(likeFileNames.size());
+    	int num = Integer.valueOf(lastFileName.substring(lastFileName.lastIndexOf("("), lastFileName.lastIndexOf(")")));
+    	System.out.println("new file name:---"+fileName + "(" + num + ")");
+		return fileName + "(" + num + ")";
+	}
+
+	@Override
+	public boolean existed(Path path) {
+    	return Files.exists(path);
+	}
 
     @Override
     public Stream<Path> loadAll() {
@@ -88,4 +113,6 @@ public class FileSystemStorageService implements StorageService {
             throw new StorageException("Could not initialize storage", e);
         }
     }
+
+	
 }
